pyproject.toml
[project]
name = "comicarr"
version = "0.6.0"
description = "Automated Comic Book Manager"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "APScheduler>=3.6.3",
    "a2wsgi>=1.7",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.8.2",
    "cfscrape>=2.0.8",
    "cheroot>=8.2.1",
    "CherryPy>=18.5.0",
    "cryptography>=46.0.5",
    "fastapi>=0.115.0",
    "feedparser>=5.2.1",
    "Mako>=1.1.0",
    "mokkari>=3.0.0",
    "natsort>=3.5.2",
    "packaging>=22.0",
    "Pillow>=10.1",
    "portend>=2.6",
    "pycryptodome>=3.19.0",
    "PyJWT>=2.8.0",
    "pystun>=0.1.0",
    "python-multipart>=0.0.9",
    "pytz>=2019.3",
    "rarfile>=4.0",
    "requests[socks]>=2.22.0",
    "simplejson>=3.17.0",
    "SQLAlchemy>=2.0",
    "sse-starlette>=3.0.0",
    "tenacity>=8.2.3",
    "tzlocal>=2.0.0",
    "urllib3<2",
    "user_agent2>=2021.12.11",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
mysql = ["mysqlclient>=2.1"]
mysql-pure = ["PyMySQL>=1.0"]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-asyncio>=0.23.0",
    "responses>=0.25.0",
    "freezegun>=1.4.0",
    "factory-boy>=3.3.0",
    "hypothesis>=6.100.0",
    "ruff>=0.15.7",
]

[tool.setuptools.packages.find]
include = ["comicarr*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["comicarr"]
omit = [
    "comicarr/rtorrent_test_client.py",
    "comicarr/dependency_check.py",
    "lib/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true

[tool.coverage.html]
directory = "coverage_html"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "E701",  # multiple statements on one line (inherited)
    "E721",  # type comparison using == instead of is (inherited)
    "E722",  # bare except (inherited — to be fixed incrementally)
    "F821",  # undefined name (inherited — needs careful review)
    "B005",  # strip with multi-char (inherited)
    "B007",  # unused loop variable (inherited)
    "B008",  # do not perform function calls in argument defaults
    "B015",  # pointless comparison (inherited)
    "B018",  # useless expression (inherited)
    "B034",  # re.sub without flags (inherited)
    "B904",  # raise without from (inherited)
    "C901",  # too complex
    "F501",  # %-format invalid format string (inherited)
    "F507",  # %-format mismatch (inherited)
    "F509",  # %-format mismatch (inherited)
    "F601",  # membership test with list (inherited)
    "F811",  # redefined unused name (inherited)
    "F823",  # undefined local (inherited)
    "W191",  # indentation contains tabs (inherited)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B", "C4"]
"comicarr/__init__.py" = ["F401"]  # re-exports used by other modules
"comicarr/helpers.py" = ["E402"]   # conditional import at line 5189
"comicarr/rsscheck.py" = ["F401"]  # urllib3 import for side effects
